backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pyodbc==5.1.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
openai==1.51.0
azure-identity==1.19.0
azure-storage-blob==12.23.0
azure-search-documents==11.6.0b8
redis==5.1.0
langchain==0.3.3
langchain-openai==0.2.3
respx==0.21.1
pytest==8.3.3
pytest-asyncio==0.24.0
email-validator==2.2.0
cryptography==43.0.3
